py_version/pyproject.toml
[project]
name = "lateblightpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4",
    "faicons>=0.2.2",
    "glob2",
    "icons",
    "ipyleaflet>=0.20.0",
    "langchain>=1.1.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-experimental",
    "langchain-google-genai",
    "langchain-openai>=1.1.0",
    "langchain-tavily>=0.2.13",
    "langgraph>=0.6.6",
    "langgraph-prebuilt",
    "langgraph-swarm>=0.1.0",
    "langsmith",
    "matplotlib>=3.10.7",
    "numpy>=1.26.0",
    "openpyxl",
    "pandas>=2.3.3",
    "python-dotenv",
    "pyrefly>=0.29.1",
    "requests",
    "scikit-learn",
    "seaborn",
    "shiny>=1.4.0",
    "shiny-calendar>=0.2.0",
    "shinyswatch>=0.9.0",
    "shinywidgets>=0.7.0",
    "statsmodels",
    "tabulate",
    "tokenizers",
    "xlrd",
]
